revamp working code
