+++ app/core/routes.py
@@ -11,82 +11,9 @@ from app import db#, logger
 from app.core.forms import PurchaseForm, SalesForm
 from app.models import User, PurchaseAgent, SaleAgent, Variety, Roles, Purchase, Sale
 from app import utilities
+from app.utilities import write_to_db
 from config import Config
 
-
-def roles_required(roles):
-    def wrapper(fn):
-        @wraps(fn)
-        def decorated_view(*args, **kwargs):
-            if not current_user.is_authenticated:
-                #return login_manager.unauthorized()
-                return False
-            urole = current_user.get_role()
-            if urole is None or urole not in roles:
-                #return login_manager.unauthorized()
-                return False
-            return fn(*args, **kwargs)
-        return decorated_view
-    return wrapper
-
-
-def write_to_db(data, db_action):
-    try:
-        if db_action is not None:
-            db_action(data)
-        else:
-            db.session.add(data)
-        db.session.commit()
-        return 200, "Success"
-    except Exception as e:
-        #logger.error(e)
-        db.session.rollback()
-        return 500, e
-
-"""
-def user_form_handler(form, action):
-    if action == "add":
-        user = User(username=form.username.data,
-                    email=form.email.data,
-                    role_id=form.roles.data)
-        user.set_password(form.password.data)
-        return write_to_db(user, db.session.add)
-    elif action == "delete":
-        user = User.query.filter_by(form.id.data).first()
-        if user is None:
-            abort(404)
-        return write_to_db(user, db.session.delete)
-    elif action == "update":
-        user = User(email=form.email.data,
-                    role_id=form.roles.data,
-                    active=form.active.data)
-        return write_to_db(user, db.session.add)
-    return 500, Exception(f"Invalid action: {action}")
-
-
-def agent_form_handler(form, action):
-    agent_type_2_model = {
-        "1": PurchaseAgent,
-        "2": SaleAgent
-    }
-    #logger.info("data is coming")
-    cur_model = agent_type_2_model.get(form.agent_type.data)
-    if cur_model is None:
-        abort(500)
-
-    agent = cur_model(
-        name=form.name.data,
-        email=form.email.data,
-        mobile=form.mobile.data,
-        address=form.address.data
-    )
-    return write_to_db(agent)
-
-
-def variety_form_handler(form, action):
-    variety = Variety(name=form.name.data)
-    return write_to_db(variety)
-"""
 
 def purchase_form_handler(form):
     purchase_data = Purchase(
@@ -118,50 +45,8 @@ def sale_form_handler(form):
     )
     return write_to_db(sale_data)
 
-"""
-def handle_form(form, form_type, action):
-    if form is None or action is None:
-        abort(500)
 
-    handler = get_form_handler(form_type)
-    if handler is None:
-        abort(500)
-    return handler(form, action)
-
-
-def get_form_handler(form_type):
-    form_2_handler = {
-        'user': user_form_handler,
-        'agent': agent_form_handler,
-        'variety': variety_form_handler
-    }
-    return form_2_handler.get(form_type)
-
-
-def is_form_support_action(form_type, action):
-    form_type_2_actions = {
-        'user': ['add', 'delete', 'update'],
-        'agent': ['add', 'delete', 'update'],
-        'variety': ['add', 'delete']
-    }
-    return form_type_2_actions.get(form_type) is not None and \
-           action in form_type_2_actions.get(form_type)
-
-
-def get_form_for_type(form_type):
-    form_dict = {
-        'user': RegistrationForm,
-        'agent': AgentForm,
-        'variety': VarietyForm
-    }
-    form = form_dict.get(form_type)
-    if form is None:
-        abort(404)
-    return form
-"""
-
-
-@bp.route('/home')
+@bp.route('/home', methods=['GET'])
 @login_required
 def home():
     generic_data = {
@@ -169,10 +54,14 @@ def home():
         "heading": "Home"
     }
     print(current_user.id)
-    return render_template("home.html", data=generic_data)
+    message = request.args.get('message')
+    if message is not None:
+        flash(message)
+    return render_template("core/home.html", data=generic_data)
 
 
 @bp.route('/purchase', methods=['GET', 'POST'])
+@utilities.roles_required([Config.ADMINISTRATOR_STR])
 def purchase():
     generic_data = {
         "title": "Purchase",
@@ -190,10 +79,11 @@ def purchase():
         #logger.info("Purchase inserted Successfully")
         return redirect(url_for('core.purchase'))
 
-    return render_template("purchase.html", data=generic_data, form=form)
+    return render_template("core/purchase.html", data=generic_data, form=form)
 
 
 @bp.route('/sales', methods=['GET', 'POST'])
+@utilities.roles_required([Config.ADMINISTRATOR_STR])
 def sales():
     generic_data = {
         "title": "Sales",
@@ -211,4 +101,4 @@ def sales():
         flash("Sale inserted Successfully")
         return redirect(url_for('core.sales'))
 
-    return render_template("sales.html", data=generic_data, form=form)
+    return render_template("core/sales.html", data=generic_data, form=form)
