+++ app/templates/admin/add_variety.html
