--- app/templates/agent.html
@@ -1,19 +1,0 @@
-<html>
-	<head>
-		<title> Broker </title>
-		<h1> Broker </h1>	
-	</head>
-	
-	<body>
-		<form name="broker" id="broker" method="post" action="" onsubmit=validate()>
-			
-			<p> Name </p> <input type="text" name="bname" id="bname" required/>
-			
-			<p> Mobile </p> <input type="text" name="bmobile" id="bmobile" required/>
-			
-			<p> Address </p> <textarea name='baddress' id='baddress' style="width:20%;height:10%;"> </textarea>
-			
-			<button type="submit" name="purchaseSubmit"> Submit </button>
-		</form>
-	</body>
-</html>
