+++ app/templates/core/sales.html
