+++ app/templates/errors/404.html
@@ -2,5 +2,5 @@
 
 {% block content %}
     <h1>File Not Found</h1>
-    <p><a href="{{ url_for('home') }}">Back</a></p>
+    <p><a href="{{ url_for('core.home') }}">Back</a></p>
 {% endblock %}
