--- login/templates/home.html
@@ -1,39 +1,0 @@
-<html>
-	<head>
-		<title> Home </title>
-	
-		<style>
-			.errorMessage {
-				color: green;
-				-webkit-animation: fadeinout 4s linear forwards;
-				animation: fadeinout 4s linear forwards;
-				opacity: 0;
-			}
-
-			@-webkit-keyframes fadeinout {
-				50% { opacity: 1; }
-			}
-
-			@keyframes fadeinout {
-				50% { opacity: 1; }
-			}
-	</style>
-	
-	</head>
-	<h1> Hello </h1>
-	<body>
-		<div id="errorMessage" class="errorMessage">
-			{{errorMessage}}
-		</div>
-		
-		<div class="btn-group">
-			<a href="/purchase">
-				<button type="submit" name="purchaseGet" class="btn btn-primary"> Purchase </button>
-			</a>
-			
-			<a href="/sales">
-				<button type="submit" name="salesGet" class="btn btn-primary"> Sales </button>
-			</a>			
-		</div>
-	</body>
-</html>
